requirements.txt
alembic==1.8.1
fastapi==0.88.0
passlib==1.7.4
pydantic==1.10.2
python-dotenv==0.21.0
SQLAlchemy==1.4.41
starlette==0.22.0
tenacity==8.1.0
sqladmin~=0.8.0